Updated resource ACL
After installation exception being thrown when trying to access store configuration:

```ssh
Invalid XML in file {...}/bitexpert/magento2-force-customer-login/etc/adminhtml/system.xml:
Element 'resource': [facet 'pattern'] The value 'bitExpert::ForceCustomerLogin' is not accepted by the pattern '([A-Z]+[a-zA-Z0-9]{1,}){1,}_[A-Z]+[A-Z0-9a-z]{1,}::[A-Za-z_0-9]{1,}'. Line: 11
Element 'resource': 'bitExpert::ForceCustomerLogin' is not a valid value of the atomic type 'typeAclResourceId'. Line: 11
```

Appears to be due to incorrect value for resource